+++ packages/server/src/session/session.module.ts
@@ -2,10 +2,11 @@ import { Module } from '@nestjs/common';
 import { SessionService } from './session.service';
 import { SessionController } from './session.controller';
 import { SessionDAO } from './session.dao';
-import { IDGenerator } from 'src/id.generator';
+import { UtilsModule } from '../utils/utils.module';
 
 @Module({
-  providers: [SessionService, SessionDAO, IDGenerator],
+  providers: [SessionService, SessionDAO],
   controllers: [SessionController],
+  imports: [UtilsModule],
 })
 export class SessionModule {}
