+++ packages/server/src/utils/utils.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { IDGenerator } from './id.generator';
+
+@Module({
+  providers: [IDGenerator],
+  exports: [IDGenerator]
+})
+export class UtilsModule {}
